% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationaritytest.R
\name{stationarity.tests}
\alias{stationarity.tests}
\title{Stationarity test}
\usage{
stationarity.tests(x, l)
}
\arguments{
\item{x}{A time series vector with missing data omitted or removed}

\item{l}{Number of lags for Ljung-Box, Box-Pierce, and ADF Tests}
}
\description{
This function runs and neatly presents stationarity tests on a series
}
\examples{
stationarity.tests(dataframe$vector, l = 5)
}
\keyword{stationarity}
\keyword{tests}
